+++ ydb/apps/ydb/CHANGELOG.md
@@ -1,3 +1,4 @@
+* The `ydb admin cluster state fetch` command was renamed to the `ydb admin cluster diagnostics collect`.
 * Added a new `--no-sanitize` option of the `ydb admin cluster state fetch` command. The new option disable sanitization and preserve sensitive data in the output.
 * Added `snapshot-ro` and `snapshot-rw` transaction modes to `--tx-mode` option of the `ydb table query execute` command.
 * Added `NO_COLOR` environment variable support to disable ANSI colors in YDB CLI (no-color.org)